Fixed issue 36902: Update.database+configScript fails under some circumstances
The problem was that update.database task fails under some circumstances if a model change
is defined in a configScript file. The problem has been found trying to update the database
in an instance with a columnSizeChange defined in the configScript file.

The problem was introduced in these improvements of DBSM sources (0036137). Before these
changesets on update.database task, the ModelChanges had been applied before and after the
platform.alterTables method was invoked. After these improvements were introduced into the
distribution, the ModelChanges had only applied after the platform.alterTables method was
invoked. (Only once)

The problem has been resolved by take into account this situation. Now the model changes are
applied (Data changes too) before platform.alterTables method is invoked in order to ensures
that when any table is recreated, all the model changes are applied properly.