Fixed issue 36611: Selection is lost when an error is raised removing a record
Backout changeset a6d788e3208b because some problems appears in the common case (succesful deletion). It is decided to fix the most uncommon case (failed deletion) by recovering the selection when the deletion fails.

As mentioned in the original fix, when were removing a row and an error was raising (e.g. Deleting a already processed "Good Shipment" is stopped by a trigger) the selection of the record was lost. This generates an inconsistent grid state that creates all
the reported problems.

Now the problem is fixed by take into account the described situation recovering the selection when the deletion fails. Now the selection is maintained if the record can't be removed properly.