Fixes issue 36922: Default values are not selected in Return Material Receipt window if the organization does not have an on hand warehouse defined
A NPE is fired if Warehouse is not selected, because was trying to access a method
of null object.

To fix that is avoided a NPE, only creating a OBDAL object if a warehouse is selected.

--HG--
extra : rebase_source : 5dcc8e9f3f8e93a08695f7e0a8f03bfa47dd5233