Fixes issue 36878: [Valued Stock Report]: Wrong Actual Cost and Valuation if the costing rule is validated with warehouse dimension.
Incorrect computation was made because costs related to transactions were not
grouped by warehouse and more than one cost could be showed with the cost of the
current line related to it. To fix this issue were took into account that if the
costing rule has the warehouse dimension activated, costs will be related to them
and calculated taking into account the transaction's warehouse. If the costing rule
hasn't the Warehouse dimension activated, then cost doesn't has the warehouse column
filled. Both cases are taken into account to a correct computation of costs.

Also was fixed when Consolidated warehouse option is choosed.