Fixes issue 36889: Wrong stock shown if Generate Aggregated Data Background process is launched with two costing rules for same organization
Wrong stock shown if Generate Aggregated Data Background process is launched with
two costing rules for same organization.

Problem was that conditions in getCostingRules() method had conditions that retrieved
more than one costing rule for the same period because an incorrect condition when
validated if starting and ending dates are null.

To fix the problem was restructured the query to combine the three conditions in just two
conditions that includes the third one (period is completely included inside the costing
rule range) and avoid return more than one costing rule for period is processed.

Also, was replaced cr.organization.id in (:org) for cr.organization.id = :org) as :org
parameter will be only one organization and not a list of organizations.