Fixed bug 36928: Emails sent by alerts contains information about other clients
The problem was that emails sent by alerts contains information about other entities. For example, when sending
an alert by email within the 'White Valley Group Admin' client, it also sends information about 'F&B Group Admin'
client. The generation of the body of the email was filtering by organization but was not filtering by client too.

The problem has been resolved by take into account the client info properly. Now the data used in the generation
of the email is filtering properly.