Fixes issue 36969: Purchase order created from Requisitions created with 0 price
Purchase order lines created from Requisitions are created with 0 price if price
list includes taxes and a discount is defined.

When M_REQUISITION_CREATEPO process is executed it creates the order and its lines.
Then it call the C_ORDER_POST1, to process the created order and it calls the
M_PROMOTION_CALCULATE to recalculate prices and quantities if discounts, and in
this case it uses the grosspricestd of lines to recalculates the gross_unit_price
and line_gross_amount if price list includes taxes:

if (v_taxIncluded = 'Y') then
update c_orderline set
    gross_unit_price = grosspricestd,
    line_gross_amount = round(grosspricestd * qtyordered, v_stdPrecision)
where c_orderline_id = Cur_Order.id;

The process was failling because order lines were created without a GROSSPRICESTD
column defined, and it was taking 0 instead of the right price. Due that, when the
values of the lines were recalculated to apply discounts it reset all values to ZERO
and order ends with incorrect totals.

To avoid that, the GROSSPRICESTD is defined when the lines are created.