Fixed bug 36886:install.source finishes with success even if an error is raised
The problem is that install.source task could finished with success even if a view cannot be created. If there is a
problem with the definition of a view (i.e. the view references a table that does not exist), the install.source
task will finish with a success message, even though error logs will be written in console output. If the
update.database task is executed instead of install.source, the task will end with error.

Now the problem has been resolved by take into account the same errors/problems on install.source task than on
update.database. The following methods now are managed in order to retrieve the potential errors and stop the
task if and error is raised:
- platform.createTables(..)
- platform.createAllFKs(..)
- platform.enableAllTriggers(..)

Now on install.source taks the errors are handled properly.