Fixed issue 36645: combo price is not correct: wrong total amount is calculated
The solution is to allow several promotions with the same ruleId in
the same line.

Additionally, there is new logic to merge promotions with the same
ruleId. The promotions are not merged if the quantity reserved is
higher than the line quantity.

It’s necessary to keep the old functionality for two families combo.

Changed price adjusment rule implementation to work as other rules.

--HG--
extra : rebase_source : 30155f21923eed349e6a98dad5f65cf516662ea3