Fixes issue 36693:Std precision can be higher than the currency's real precision
In Currency window, Standard precision can be higher than the currency's real
precision (commonly 2). The Standard Precision is used to round all the amounts
(not prices) of the documents created in the backend.

Was created a new callout used to validate the currency standard precision. If
the new precision is higher than specified currency precision in ISO 4217 Currency
codes, then a warning is showed to user.

Also, if the new precision is negative then an error message is shown and it is
set as the default standard precision: 2.