Fixed issue 36974: [SERQA 3316] It is possible to bypass the overpayments with Pay Open tickets
The solution is to check first if the last payment has an overpayment and if
the payment method has defined a limit for it.

Since right now it’s not possible to add more payment once the user has pay it
completely, this is a valid solution.

This solution has forced to take the overpayment to a new function called
checkValidOverpayment and change the old function checkValidCashOverpayment
(now renamed to CheckValidCashChange).