Fixed issue 37051: Rich text area fields causes visibility problems in the grid
The problem was that the rich text area fields generated visibility problems in the grid under some circumstances.
(e.g. when a rich text area field has a carriage return or html tags such as div, br,etc.). A common usecase could
be when a text is copied from a website.

This visibility problems occurred because in form view this type of fields are visualized as a rich text editor and
html tags are supported but in grid view are visualized as a normal field in the grid and this html tags are not
properly handled by SC.

In order to avoid this grid visibility problems, when showing a rich text area field in grid view, now it is escaped
the html code. Visibility in form view is not changed. Now the problem has been fixed by taking account this situation
in getGridFieldProperties method of the RichTextUIDefinition reference.