Fixed issue 36838: multiple requests to load product image are performed
The problem was that multiple requests were performed when it shouldn't. When a product had an image, multiple
requests were being performed when the product was beeing displayed in form view and a refresh was being done.
Another problem discovered in the research of the issue was that a request was being performed when return from
FormView to the GridView. This request should not be performed because images are not showed in the grid view.

Both problems have been fixed by take into account when a request should be performed to retrieve an image. When
the image is loaded previously, the setValue of the ob-formitem-image avoids to performed an unnecessary request
to the server.