Fixes issue 36466: "Copy from order" refactor to Process Definition
Currently the process "Copy From Order" is defined as a manual process and does
not make use of the benefits of Process Definition. The current process logic
was migrated to a new Process Definition, using the Pick and Edit (PE) pattern
in those windows that currently use it.

Was created a new Pick and Execute window. Defined a new tab with the same fields
that the old process.

Was created a new Process Definition using previously windows as parameter and
updated all the buttons that were using the "Create From Order" process to use
the new process definition "Create From Orders".

Was created a new action handler to process the selection of the orders. It
iterates through all the selected orders and add their lines to the order that
executes the process.