Fixed issue 36605: unneeded datasource requests opening alert management window
The problem was when Alert Management window was opened, 10 requests to datasource were performed.
In order to improve this situation, several actions have been done. Firts, the 4 calls of the datasource
definition are removed. Now, the datasource is defined statically. Thisthat Alert Management window is
not working good. Also, two of the duplicate request are removed.

Besides, when the Alert Management window is closed and the four grids are removed, now the datasources
are destroyed too in order to avoid having several instances for the same datasources every time that
Alerts windowd is reopened. Now, every grid removes their own datasource.