Fixed issue 37180: Implements the functionality to generate separate invoices for return and sales tickets
If terminal type is set to group orders in one invoice, a new check is shown.

The use of this check Separate Invoice for Return allows to the client to
generate one invoice for all sales order and another one for returns.

--HG--
extra : rebase_source : 79693f0df407ebcb354392b62049fd6eac7dba9a