[jsimport cleanup] Removed unused searchs,calendar,messages imports.
That file is not a report but small html 'popup' on print button of
Business Partner window.
It does not use any 2.50 style selector or calendar or the message system.
So remove those unused imports. Note that removed DynamicJS import also had
a wrong path so was triggering 404 not found error before also.