Fixed issue 37229: Incremental refresh is not working deactivating products in Discount & Promotions window
During the full refresh, we can avoid those records that are deactivated in
Product sub-tab from Discount and Promotion window. During an incremental
refresh, we want to search for all records (deactivated or not) in product
sub-tab.

Implement logic to add Product filter from Discount and Promotions window
only during a full refresh.