Related to issue 35741: Refactored 'Delete Lines' process
Refactored the delete lines process in order to have a better performance and be able to save correctly the deleted lines.

Now, there's a new API that must always be called in order to remove a single or multiple lines. The new function is the function 'deleteLinesFromOrder' located in the order. Using this new API is possible to control the asynchronous flows and we are able to do the delete lines process executing the 'calculateReceipt' process only once, getting a notable performance improvement.

The lines are now marked to be deleted, and in the final callback are removed at once.

The old delete line process is still alive, but a warning is added in the console telling that the flow is deprecated and that when deleting a line, the related services won't also be removed (they will be updated during the final quantity rules process, but will not be stored in the deleted lines array and won't be available is pressing the undo button).

--HG--
extra : rebase_source : f420f835643f8960dcb0bfcd2aacc872172bc74f