fixes bug 37255: Tree in subtab is closed after pressing 'Enter' key to filter
Being on a child tab, the fix for issue #37181 was not complete because it was comparing the entered criteria without including the criterion with the parent record id. Thus when comparing the current criteria and the entered criteria they were always different.

Now we are completing the fix by converting the entered criteria in order to include the filtering criterion by parent id.

--HG--
extra : rebase_source : a4ed70376e40f48a383b31db5cc820c8740f540d