Fixes issue 35838: Improved the performance in Alert Management window
When you had several alerts (the problem was reported with +20000 alerts) and the Alert Management
window was opening, there was a performance problem: The alerts weren't show, java was consuming
the memory,etc.

Several refactors and improvements have been done in order to improve the performance in Alert
Management window:

- The getWhereAndFilterClause method was performing twice per alert status. It is removed the
invocation from the fetch method in ADAlertDatasourceService because it is done in the fetch method
of the super class.
- It is not neccesary to retrieves the full alert object when only the alert ID is needed. The same
for AlertRule.
- Now the status of the alert is take into account in the queries in order to retrieves from the DB
only the alerts that should be needed.
- StringBuilder is used instead of String when concatenate several IDs.

Now the performance of the window is improvement a lot. For example, having more than 20000 alerts,
it has been reduce the time from 40.1 seconds to 0.6 seconds (Times in DataSourceServlet.doFetch)
showing to the user the same information.

--HG--
extra : rebase_source : cfa8d9f12ecdd5020dfaffab0992590525715b1e