Fixed issue 37214: Improved performance in SQL Query manual window
It was possible to produce a server crash executing on the sql query window a query returning
thousands of records. An OutOfMemoryError error is raised although only 20 records were showed
to the user because in fact, all the records were retrieving from the db instead of the 20 needed
records.

In order to fix the performance problem, now the queries are executed as a Native SQL with OFFSET
(org.hibernate.SQLQuery.setFirstResult) and LIMIT (org.hibernate.SQLQuery.setMaxResults). Now the
performance problem has been fixed.