Fixed issue 37278: Deactivated roles are displayed in the user profile widget
The problem was caused by the "Improve scalability of the login process" project. When improving the performance
of the query that return the complete role list for the current user, the OBQuery instance used to generate it was
replaced with a Query instance. The first includes by default the filter by is active.

To fix the problem this filter must be added manually in the query. Now the problem has been resolved by take into
account this filter.