Fixed issue 37289: Deactivated orgs are displayed in the user profile widget
The problem was caused by the "Improve scalability of the login process" project. When improving the performance
of the query that return the complete organizations list for the current user, the OBQuery instance used to generate it was
replaced with a Query instance. The first includes by default the filter by is active.

To fix the problem this filter must be added manually in the query. Now the problem has been resolved by take into
account this filter. Besides the same problem is detected in other query related with deactivated warehouses so it is fixed too.