Fixes issue 37302: Process Price Difference Correction fails with n.p.e. or create wrong Cost Adjustment Documents when it works
Price Difference Adjustment process is invoked for legal organizations.

When invoked from manual process, the legal organization(s) of selected
organization is retrieved, the process is invoked for that legal organization
in order to the Cost Adjustment Document be created with the legal organization
in the header, and lines for transaction belonging to organizations under it.

When invoked from the background process, the process is invoked for legal
organizations under organization "*".

When invoked from the Costing Background process, the legal organization of the
transaction organization is retrieved, and is used to invoke the process.

--HG--
extra : rebase_source : 759f306e994574911508b7241763d8d989980097