Fixes issue 37317: Costing Background Process stopped if the same product is produced and consumed in different production run of same work effort
If the product is not consumed by the same work effort, the costing server will
create/update the cost, otherwise, the cost is updated, to be taken into account
for the production run that consumes it.

This fix is a contribution of Support Team.

--HG--
extra : rebase_source : ea957c3ef037bbb887374ffa42a45141b918be40