Related to issue 36466: Fixed preferences conflicts with AWO
To avoid possible conflicts with other tests using the UomManagement preference
the Copy From Orders tests related with the AUM functionality were adapted to save
the status of the preference at the start of the tests, and restore it after the
all the tests are executed.

Is taken into account if the preference exists in the database, and in case the
preference exists also is restored its value. This way, at the tests ending, it
left the preference status as expected by other tests.

Also was created a new class to manage all the actions related with the preference.
Some of existing methods were moved to it, and other were created.

--HG--
extra : rebase_source : f87fb66b1e47035e974a8626b2c42969963d45cc