Fixed issue 37044: [SER QA 3365] Wrong ticket amount when action made before ticket total amount calculated
The solution for this issue is to show a blocking screen during the calculation
of the ticket. Exists two ways to activate the new functionality:
  * At POS Terminal Type window, there is a new check called “Block screen
     when processing ticket”. By deafult is uncheked
  * At preference level, called "Block screen when processing ticket on slow
     devices". By default, is set to ‘Y’

If the conditions are met, a screen is shown during the calculation of the receipt
and the WebPOS is blocked. If one of these conditions are met, the blocking
screen is shown:
  * If the value Block screen when processing ticket” is checked
  * If the preference “Block screen when processing ticket on slow devices” is
     set to ‘Y’ and the benchmark is less than 1000.

When the calculateReceipt is started, a transparent blocking screen is show to
block the webPOS. If the receipt calculation takes more than one second, a
semitransparent screen with a popupmessage is show to indicate the calculation
is ongoing. At the end of the calculateReceipt, all blocking screen are removed.

--HG--
extra : rebase_source : 6c2e372ff8636e7fadabfaff9bce7145563b60cf