Fixes issue 34628: Manage variants can return unexpected values
When filtering concurrently in Manage Variants P&E, it was possible to obtain
incorrect values. This is caused because ManageVariantsDS is not thread safe it
is used as singleton but it has state variables that can be messed up by
different threads.

The global state variables has been removed and now their values are contained
in a new private class. A new instance of this class is created for every execution
of the process, so the values are not mixed again.

The getData and readCriteria methods have been modified to make use of this new
class.

--HG--
extra : rebase_source : d267a04587c6f6b3424fd35d034307c956e26ea7