Fixes issue 37580: "division by zero" error when book sales order with BOM products
When closing a sales order with a BOM product a division by zeror error was raise
while the process tries to calculate the v_PriceActual variable.

A validation was added to only calculate the value when the qtyordered is
different from cero, if not the v_PriceActual value is cero.