Fixed issue 33307: Role with access to current store will be selected when users try to login in web POS
If a user is login in web POS, and it does not have a default role for web POS, an additional logic will be executed to select a role which could allow to the user to login into the terminal which user is trying to use.
This logic only will be executed when
  1. User does not have default role for web POS
  2. User have several roles assigned
  3. At least on of this roles assigned is marked to use web POS (Has acces to the form web POS)

New logic will iterate roles assigned to the current users marked with access to web POS, and will select the first role which have access to the organization where the current terminal belongs.
This logic can work randomly when a user can use several roles allowed to use web POS with access to the same organization. In this case first role found will be selected.