Fixed issue 37593: Avoid unnecessary services checks.
A new property 'hasServices' has been created in the order, which tells if an order contains any service that is related to a product. Each time a quantity rule or any check that affects to services should be executed, this property is checked and the check is done only for those tickets that contain a service.

This property is set when creating services lines, and is removed when deleting the services lines from the ticket.

Additionally, the duplicated code executed to check if services can be returned has been improved to be better in performance and more readable and also has been moved to the order.js, creating a function to avoid the duplicated code and having the business logic in the correct file.

--HG--
extra : rebase_source : d7b4c06886689e2cf835df7d990962016dca15bb