Fixes issue 37589: remove setTimeout when clicking in services button, and add a return true so select line button is not executed
The button of services is inside the ticket line that is also a button,
and inside button executes before, so first open the services window and after the edit,
to change this order the services windows is opened with a setTimeout.

Now changed to only execute service button, and with a return avoid to execute the select line button,
so the setTimeout is no longer needed.

This change is needed for the throttle, since the timeout runs when the component is already destroyed and fail.

--HG--
extra : rebase_source : fe209a480f6831906667d2801f8cf796d451a9f7