Fixes issue 37643: OBScriptEngine can be used to evaluate javascript scripts
There were four classes that were creating a new instance of ScriptEngine each time they needed to evaluate a javascript script. This was not the better way
to do it, since creating a ScriptEngine instance can be expensive, and that class is thread safety if properly used (i.e. by setting the bindings to each
evaluation instead of to the script engine itself).

A new singleton class has been added, OBScriptEngine, that is now used by all the classes that need to evaluate javascript scripts.