Fixes Issue 30643:Overly huge request size in AddPaymentDisplayLogicActionHandle
Before sending the context, the grids with the information about orders and invoices,
gl items and credit used are removed from it.
This data is not used for calculating the display or read only logic of the rest of the parameters
and sending the grid can have an impact in the performance of this process.

--HG--
extra : rebase_source : fe505fcd813a3571806106b463e9c55c67d851bb