Fixes issue 37677: Dont check if ad_context_info table exists before creating it
Now, if the pool is used, the query to check if the ad_context_info table already exists will not be done. This check can be skipped because
if the pool is used, the initDB method will only be called when a new physical connection is created, and in that case the temporary table is
garanteed not to exist.

This pool instance check is needed to keep supporting using pools other than org.openbravo.apachejdbcconnectionpool.JdbcExternalConnectionPool (or none at al),
but in those cases the query to check if the ad_context_info table exists will still be performed.