Fixes issue 37681: Dont create ad_context_info if audit trail not active
Records are only read/written from the ad_context_info if audit trail is enabled (if any ad_table is fully audited). If audit trail is disabled when
Tomcat is started, there is no need to create the ad_context_info table each time that a new connection is created. From now on it will no be possible
to enable the audit trail while Tomcat is active, enabling it will require to rebuild Openbravo and to restart Tomcat. Now:
- The Update Audit Trail Infrastructure has been removed
- When a new table is set to be audited in the AD, SessionInfo.setAuditTrailActive(true) will not be invoked, so that the isAuditTrailActive flag can
only be set to true when Tomcat starts.