Fixed issue 37684: Filter on quick launch menu is not accent sensitive
After the changes introduced in the "Improve scalability of the login process" project on issue #35553, the quick launch menu no longer used the specific datasource. The datasource has a fieldMatchesFilter method that allow searchs not distinguish between accent or non-accent words. Since the changes were introduced, the data is filter in the client instead of in the datasource and for this reason the fieldMatchesFilter method is not executed.

To fix the problem two main actions have been done:
- Small clean code moving the functionality of fieldMatchesFilter method to ob-utilities in order to avoid duplicated code.
- The getQuickMenuItems method must be save the identifiers without accents using the method previosuly created in ob-utilities. It has been done the same with the value returned in the getPickListFilterCriteria method.

Removing the accents in the items of the filter in ob-quick-launch.js the problem has been resolved. The accents are now ignored in Quick launch Menu' suggestion combobox.