Fixes Issue 34496. This class was original developed as a singleton. However, as it evolved, this design was changed and now it behaves as an object.
Now the information of the object should be created and destroyed
when a new instance of this object is created.

This is done in CostAdjustmentProcess using injection.

That is why the definition of ApplicationScoped does not make
sense any more, it must be Dependant.

--HG--
extra : rebase_source : 68706639ab861ec6bbaabfbfdbbd75327ee1e243