Fixes Issue 37764. Fixes performance in Return Material Receipt P&E
In view m_rm_receipt_pick_edit a new column has been added with the
Business Partner from the Order.

This new column is used in the filter of the Return Material P&E Tab.
By doing so, additional cross joins with OrderLine and Order Tables
are avoided.
Also, the planner in Postgres is able to properly use the Business
Partner index.