Fixes issue 37724: Error when cancel and replace a sales order of zero amount
When an order with grand total of ZERO is created and booked, no Payment Plan is
created for it, so when she is cancelled and replaced the new created orders haven't
payment plans neither.

So, when the Cancell and Replace process is executed on orders of this kind, the
it was failing trying to get get the payment schedule detail of the oldOrder and
an exception was thrown.

To fix this issue, only will be created Payments to orders with grand total different
than ZERO to avoid the exception and keep the current behaviour for this kind of orders