Fixes issue 37451: Fixes filtering of fks that dont use ids as key
The problem was that the foreign keys that did not use and id column as key (i.e. ad_country.ad_language) were not being properly
filtered locally in smartclient, because the criteria used the id of the referenced record, that was not available in the browser (it
was available in the query done in the back end, that is why only the client-side filtering was not working).

For instance, if the ad_language foreign key reference uses as key ad_language.ad_language instead of ad_language.ad_language_id, the criteria
will no longer be:

{
  fieldName: 'language',
  operator: 'equals',
  value: '192'
}

Now it will be:

{
  fieldName: 'language',
  operator: 'equals',
  value: 'en_US'
}

The AdvancedQueryBuilder has been adapted to support the new type of criteria.

There are also changes on ob-grid.js and ob-formitem-fk-filter to ensure that the criteria is properly displayed in the filter input.