Related with issue 37451: Adds tests
Tests have been added to check that the datasource returns properly the results using the updated criteria.

Selenium tests with the current infrastructure are not possible, since when they filter the grid, they do not
select the checkboxes of the foreign key combo boxes, they just enter the text in the filter input, and if done
that way the issue is not reproducible.