Fixes issue 37300: Draft ticket without product is printed as Return receipt
In pointofsale-print.js, when the receipt has no lines, the variable hasNegativeLines
is always taking true as its value, which caused the assigned template for
printing to be the one for returns. A validation has been added, to verify if the
receipt has lines, and only assign the return template if it is true.