Fixes issue 37823: Sales tax selected in Return To Vendor if the returned goods receipt does not have a purchase order related
A wrong sales tax was selected in Return To Vendor if the returned goods receipt
does not have a purchase order related.

The Tax field has the Tax Id value of the receipt line being copied, in case
that the good receipt line was created from an order then it will have the tax id
used on the order line. But in case the goods receipt line was created manually, as
it lines aren't related to any order lines, the tax received will be NULL, so it is
needed to get the default Tax to be used in the new generated returned line.

In the case of a Return to Vendor is needed to return the default Purchase Tax of the
product. In the case of Return from Customer is neede to use the default Sales tax
of the product. It was fixed passing this parameter taking into account the processed
order is sales transaction or not.