Fixes issue 37897: Callouts are invoked when clearing a non mandatory selector
This changeset [1] fixed a problem where a callout was being called twice when picking a value from a selector. The first request was done properly, but the
second one was done with a null value. The fix consisted in not notifying a change in selector when it was modified from a non-empty value to an empty value.

But as a result, when a non mandatory selector is emptied, callouts are no longer being invoked. The original issue is very hard to reproduce, I have not been
able to find a new way to reproduce it or fix it. To avoid the regression, now the code that was done to fix [1] will only take place if the selector is mandatory,
to make sure that it is possible to give empty values to non-mandatory selectors.

[1] https://code.openbravo.com/erp/devel/pi/rev/baf72de5b678a308c36d5946af3e143d08625eca