Related with issue 37681: Dont call to setAuditActive(false) in SL_TableAudit
The SL_TableAudit callout is invoked when changing the value of the Is Fully Audited flag in ad_table. When the user set this flag to false to the only table that was
being audited, the callout was invoking SessionInfo.setAuditActive(false); to notify that there were no more tables being audited.

But this can lead to problems. Suppose:
- There is one table audited, the audit trail triggers are installed
- All connections have created the ad_context_info table because of this
- The user sets to false the Is Fully Audited flag of the audited table and invokes SessionInfo.setAuditActive(false);
- From then on new connections will not create the ad_context_info table. But the triggers are still installed in the table, because the audit trail infrastructure is not
updated until the system is rebuilt and restarted. If there is a change in data of that table and the connection in charge of doing it has not created the ad_context_info,
the change will not be commited because the triggers will fail.

--HG--
extra : rebase_source : df5bfb3a3c544feb57bd31147b6d1fb23b39fe53