Fixes issue 37025: added pagination by Id to product price
- Change the product price model to use the pagination by Id instead the old one by offset

- Change a bit the product price master data load query:
Put all possible inside the subquery
This toguther with the cange of paginate by id instead , gives half times loading many pages, in some scenarios.

Note: if separeted in two queries, it improves a lot, but in future versions of pg could give worse performance,
Also the number of parameters to the second query could be a problem.

- Fix ServiceProduct master data load query, since the ordering was not correct.