Fixes issue 37850: Stock Valuation of 0 in Product - Unit Cost tab if product is received in a warehouse created after unit cost is calculated
When a new transaction increase the product stock in a warehouse and doesn't exist
a stock valuation record to this warehouse the M_Transaction_Cost trigger can't
set the valuation cost because the stock is 0.

The flush to M_Transaction when the cost is calculated trigger the M_Update_Stock_Valuation
with the correct stock quantity and then the M_Transaction_Cost trigger can set the
valuation cost to the Stock Valuation created in the M_Transaction trigger.

--HG--
extra : rebase_source : 4b81f06040aeb21e4148268fba21d78cefc181fc