Fixes issue 37882: Cannot add a purchase order line to a goods receipt if the Alternate UOM feature is activated after the order was created.
If AUM is not enabled, when a purchase order is created it hasn't values in any of
the columns related with AUM. So, if the preference is activated after the order
creation, the ordered quantity will be used as AUMQTY and the UOM as AUM by default
when these order lines are copied to the receipt by the Create Lines From process and no
alternate units exists for the ordered product. In the case that alternate unit of
measure exists for the product the default AUM will be used as the receipt line AUM
and to calculate the operative quantity based on the conversion rate between the UOM and the AUM.

Were added to the XSQL file the return of the document type as C_DOCTYPE_ID field to use it when
the default AUM is calculated.

--HG--
extra : rebase_source : be91c060189dfed73c395c8e13e9d0f41414fddd