Fixed bug 38015: [ri] Attribute Set sometimes is not copied when multiboxing
When calculating the new attribute set instance per storage detail for boxing, the system first tries to recover any previously cloned attribute set instance for this referenced inventory.
It can find it either in the cache created in this boxing execution, or by searching in the database for an existing attribute from a previous boxing activity for the same attribute.
If a previously cloned attribute set instance is not found, it creates a new one.

The existing index has been extended to perform properly in the query to search for an existing cloned attribute set instance.

--HG--
extra : rebase_source : 61bf3808b394bf8df3220e6dace74f71c4fd0a1b