Fixes issue 37939:Cannot confirm Cancel and Replace if a Service Product related to a product was included on the Sales Order
After the new lines were created in the Cancel and Replace process the relations
between services and products were not copied from the replaced order, due that
the Confirm and Cancel process was failing.

To fix that, if exists relationships between the lines of the creplaced order they
are copied taking as reference the new order lines.

--HG--
extra : rebase_source : 290a6b276d09e3d37102ae3fe68cd6a50fb78426