Fixes issue 37960: When not stocked product is exploded before booking the order and price includes taxes, Gross Unit Price is updated to 0.
When the it is exploded a BOM line with not stock and the price list includes taxes
then the M_EXPLODEBOMNOTSTOCK function was creating lines with the gross_unit_price,
grosspricelist and line_gross_amount values, but it wasn't copying the grosspricestd
column and it was taking 0 as default value.

For this reason, when the order was booked, as the Explode process sets the
grosspricestd to 0, M_PROMOTION_CALCULATE function uses and transfers it for the
gross_unit_price column.

To fix this issue, the grosspricestd is also inserted into the exploded lines at
line insertion in M_EXPLODEBOMNOTSTOCK PL.