Fixes issue 37963: Payment in Awaiting payment status is created when an invoice cannot be voided because document cancelled is a return
A commit and close was performed at the end of the invoice processing to get the
error message from the database, but it left all previously created data into the
database just like the dummy payment when it is needed is to rollback the changes.

If at the end of processing there is some error then a rollback is performed.
Also the message is created from the pinstance object instead of retrieve the Instance
data from the database.