Fixes issue 38045: Improved performance when executing the payment report
The query to retrieve the invoices using credit in getInvoicesUsingCredit method
was restructured to improve the performance, avoiding the join with the FIN_Payment_Detail_V
view and extracting the logic of it but without do some joins with not needed tables on
this scenario. Also the new query is using the defined indexes to execute the query
faster and decrease the cost of the execution.

--HG--
extra : rebase_source : 4f74b646c88c6c980c9dcc6633b0b0b2ee31a8ba