Fixes issue 38157: Receipts originally calculated with avg and adjusted by price difference correction are readjusted if avg cost changes.
Having a Goods Receipt with no related Purchase Order but a related Purchase
Invoice, when the cost of the receipt is calculated the current average is used.
If the average cost used to calculate it is later changed by some cost adjustment
the cost of this transaction is calculated again even if the receipt was adjusted
by the Price Difference Correction process.

Now when checking if the transaction is based on the average cost take into account
if the receipts are already adjusted by the Price Difference Correction.

--HG--
extra : rebase_source : 3851f1bfe440d7bc150d56a5c8c756b711997b77