[Prepayments] Fixed issues in payments distributions for order and invoices
Fixed the remaining amount after a reverse payment that was not paying fully the ticket.
Also, when adding a payment to a ticket with a partial invoice, the payment was being priorized to fulffil the not invoiced part.
Also the code has been cleaned to avoid not necessary checks, like checking if a schedule detail has a payment detail when the query to find the details is filtering for those payments that has details.

--HG--
extra : transplant_source : %D0%0E%85%FB%5C%5B%7B1nL%E7-%82%2Br%93m%FD%40%CB